Print Files and Interiors
Exported STL and OBJ for each stage. To keep the engines from falling off, I added a few pods to act as structural connections (not realistic, just for printing). These new features are ignored in all other analyses.

Fixed the interiors again so that the volumes all line up properly with the OML. This effort was done previously for the report, but then VSP crashed and I lost all my work.